Convert long values to int. In theory the maximum integer value what comes from the addon side is 2^24 bit - 3 byte and thus avoid unnecessary number conversion. AuraCount is now struct

+++ Core/Addon/BuffStatus.cs
@@ -2,9 +2,9 @@
 {
     public class BuffStatus
     {
-        public long Value { get; private set; }
+        public int Value { get; private set; }
 
-        public BuffStatus(long value)
+        public BuffStatus(int value)
         {
             this.Value = value;
         }
